New entry point and update to example build script
This commit adds a new check for an '_start' symbol for program entry to
elf2sun entries. '_start' is an alterante entry point that calls the
main function. It was added to libp to support handle main function
returns in the future. We also provided comments for optional build
script libp object files.